titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/eposs/cut-piece-tot/data-access-cut-piece-tot/src/lib/+state/cut-piece-tot.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectCutPieceTotState } from './cut-piece-tot.reducer';

const selectCutPieceTotDetails = createSelector(
  selectCutPieceTotState,
  state => state.cutPieceTotDetails
);

const selectIsLoading = createSelector(
  selectCutPieceTotState,
  state => state.isLoading
);

const selectUpdateCutPieceTotResponses = createSelector(
  selectCutPieceTotState,
  state => state.updateCutPieceTot
);

const selectError = createSelector(
  selectCutPieceTotState,
  state => state.error
);

export const CutPieceTotSelectors = {
  selectCutPieceTotDetails,
  selectUpdateCutPieceTotResponses,
  selectIsLoading,
  selectError
};
